refactor: 구성원 관리 메뉴 폴더로 묶기 #11
